refactor(phase-4): split secretsantaext into package + features/secretsanta
Break the 1161-line secretsantaext.py god-extension into two layers:

- features/secretsanta/ — non-Discord code: SecretSantaSession dataclass,
  SecretSantaRepository (MongoDB I/O), and pure assignment algorithms
  (single-cycle backtracking with MRV + forward checking, and subgroup
  cycles).
- extensions/secretsanta/ — thin Discord mixins split by concern:
  sessions (create/participants/cancel), draws (draw/reveal/remind),
  bans (banpair/unbanpair/listbans), and buttons (join/leave).

Each subcommand is declared with sub_cmd_name on @slash_command so they
can be distributed across mixin modules while still registering under a
single /secretsanta group.
