+++ features/secretsanta/__init__.py
@@ -1,0 +1,15 @@
+from features.secretsanta.assignments import (
+    generate_assignments_with_subgroups,
+    generate_valid_assignments,
+    is_valid_assignment,
+)
+from features.secretsanta.models import SecretSantaSession
+from features.secretsanta.repository import SecretSantaRepository
+
+__all__ = [
+    "SecretSantaRepository",
+    "SecretSantaSession",
+    "generate_assignments_with_subgroups",
+    "generate_valid_assignments",
+    "is_valid_assignment",
+]
